Moved DefaultClock to DefaultClock.php file to fix PHPUnit autoload for tests
WE2-868
