+++ src/util/DateAndTime.php
@@ -78,38 +78,3 @@ final class DateAndTime
         return ((clone $date)->setTimezone(new DateTimeZone("UTC")))->format("Y-m-d H:i:s e");
     }
 }
-
-/**
- * @copyright 2022 Petr Muzikant pmuzikant@email.cz
- */
-final class DefaultClock
-{
-    private static DefaultClock $instance;
-    private DateTime $mockedClock;
-
-    public static function getInstance()
-    {
-        if (!isset(self::$instance)) {
-            self::$instance = new DefaultClock();
-        }
-        return self::$instance;
-    }
-
-    public function now(): DateTime
-    {
-        if (isset($this->mockedClock)) {
-            return $this->mockedClock;
-        }
-        return new DateTime();
-    }
-
-    public function setClock(DateTime $mockedClock): void
-    {
-        $this->mockedClock = $mockedClock;
-    }
-
-    public function resetClock(): void
-    {
-        unset($this->mockedClock);
-    }
-}
